feat(ui): implement scaffold for home page
