+++ frontend/src/app/dashboard/layout.tsx
@@ -1,0 +1,11 @@
+import type { ReactNode } from "react";
+
+import { LiveConvexProvider } from "@/features/live-convex-provider";
+
+interface LayoutProps {
+  children: ReactNode;
+}
+
+export default function Layout({ children }: LayoutProps) {
+  return <LiveConvexProvider> {children}</LiveConvexProvider>;
+}
